# strava/aspnetcore/src/IO.Swagger/Controllers/ClubsApi.cs
/*
 * Strava API v3
 *
 * The [Swagger Playground](https://developers.strava.com/playground) is the easiest way to familiarize yourself with the Strava API by submitting HTTP requests and observing the responses before you write any client code. It will show what a response will look like with different endpoints depending on the authorization scope you receive from your athletes. To use the Playground, go to https://www.strava.com/settings/api and change your “Authorization Callback Domain” to developers.strava.com. Please note, we only support Swagger 2.0. There is a known issue where you can only select one scope at a time. For more information, please check the section “client code” at https://developers.strava.com/docs.
 *
 * OpenAPI spec version: 3.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ClubsApiController : ControllerBase
    { 
        /// <summary>
        /// List Club Activities
        /// </summary>
        /// <remarks>Retrieve recent activities from members of a specific club. The authenticated athlete must belong to the requested club in order to hit this endpoint. Pagination is supported. Athlete profile visibility is respected for all activities.</remarks>
        /// <param name="id">The identifier of the club.</param>
        /// <param name="page">Page number. Defaults to 1.</param>
        /// <param name="perPage">Number of items per page. Defaults to 30.</param>
        /// <response code="200">A list of activities.</response>
        /// <response code="0">Unexpected error.</response>
        [HttpGet]
        [Route("/api/v3/clubs/{id}/activities")]
        [ValidateModelState]
        [SwaggerOperation("GetClubActivitiesById")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<SummaryActivity>), description: "A list of activities.")]
        [SwaggerResponse(statusCode: 0, type: typeof(Fault), description: "Unexpected error.")]
        public virtual IActionResult GetClubActivitiesById([FromRoute][Required]long? id, [FromQuery]int? page, [FromQuery]int? perPage)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<SummaryActivity>));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(Fault));
            string exampleJson = null;
            exampleJson = "[ \"\", \"\" ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<SummaryActivity>>(exampleJson)
                        : default(List<SummaryActivity>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// List Club Administrators
        /// </summary>
        /// <remarks>Returns a list of the administrators of a given club.</remarks>
        /// <param name="id">The identifier of the club.</param>
        /// <param name="page">Page number. Defaults to 1.</param>
        /// <param name="perPage">Number of items per page. Defaults to 30.</param>
        /// <response code="200">A list of summary athlete representations.</response>
        /// <response code="0">Unexpected error.</response>
        [HttpGet]
        [Route("/api/v3/clubs/{id}/admins")]
        [ValidateModelState]
        [SwaggerOperation("GetClubAdminsById")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<SummaryAthlete>), description: "A list of summary athlete representations.")]
        [SwaggerResponse(statusCode: 0, type: typeof(Fault), description: "Unexpected error.")]
        public virtual IActionResult GetClubAdminsById([FromRoute][Required]long? id, [FromQuery]int? page, [FromQuery]int? perPage)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<SummaryAthlete>));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(Fault));
            string exampleJson = null;
            exampleJson = "[ \"\", \"\" ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<SummaryAthlete>>(exampleJson)
                        : default(List<SummaryAthlete>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Get Club
        /// </summary>
        /// <remarks>Returns a given club using its identifier.</remarks>
        /// <param name="id">The identifier of the club.</param>
        /// <response code="200">The detailed representation of a club.</response>
        /// <response code="0">Unexpected error.</response>
        [HttpGet]
        [Route("/api/v3/clubs/{id}")]
        [ValidateModelState]
        [SwaggerOperation("GetClubById")]
        [SwaggerResponse(statusCode: 200, type: typeof(DetailedClub), description: "The detailed representation of a club.")]
        [SwaggerResponse(statusCode: 0, type: typeof(Fault), description: "Unexpected error.")]
        public virtual IActionResult GetClubById([FromRoute][Required]long? id)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(DetailedClub));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(Fault));
            string exampleJson = null;
            exampleJson = "\"\"";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<DetailedClub>(exampleJson)
                        : default(DetailedClub);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// List Club Members
        /// </summary>
        /// <remarks>Returns a list of the athletes who are members of a given club.</remarks>
        /// <param name="id">The identifier of the club.</param>
        /// <param name="page">Page number. Defaults to 1.</param>
        /// <param name="perPage">Number of items per page. Defaults to 30.</param>
        /// <response code="200">A list of summary athlete representations.</response>
        /// <response code="0">Unexpected error.</response>
        [HttpGet]
        [Route("/api/v3/clubs/{id}/members")]
        [ValidateModelState]
        [SwaggerOperation("GetClubMembersById")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<SummaryAthlete>), description: "A list of summary athlete representations.")]
        [SwaggerResponse(statusCode: 0, type: typeof(Fault), description: "Unexpected error.")]
        public virtual IActionResult GetClubMembersById([FromRoute][Required]long? id, [FromQuery]int? page, [FromQuery]int? perPage)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<SummaryAthlete>));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(Fault));
            string exampleJson = null;
            exampleJson = "[ \"\", \"\" ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<SummaryAthlete>>(exampleJson)
                        : default(List<SummaryAthlete>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// List Athlete Clubs
        /// </summary>
        /// <remarks>Returns a list of the clubs whose membership includes the authenticated athlete.</remarks>
        /// <param name="page">Page number. Defaults to 1.</param>
        /// <param name="perPage">Number of items per page. Defaults to 30.</param>
        /// <response code="200">A list of summary club representations.</response>
        /// <response code="0">Unexpected error.</response>
        [HttpGet]
        [Route("/api/v3/athlete/clubs")]
        [ValidateModelState]
        [SwaggerOperation("GetLoggedInAthleteClubs")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<SummaryClub>), description: "A list of summary club representations.")]
        [SwaggerResponse(statusCode: 0, type: typeof(Fault), description: "Unexpected error.")]
        public virtual IActionResult GetLoggedInAthleteClubs([FromQuery]int? page, [FromQuery]int? perPage)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<SummaryClub>));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(Fault));
            string exampleJson = null;
            exampleJson = "[ \"\", \"\" ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<SummaryClub>>(exampleJson)
                        : default(List<SummaryClub>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}