# strava/aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Strava API v3",
    "description" : "The [Swagger Playground](https://developers.strava.com/playground) is the easiest way to familiarize yourself with the Strava API by submitting HTTP requests and observing the responses before you write any client code. It will show what a response will look like with different endpoints depending on the authorization scope you receive from your athletes. To use the Playground, go to https://www.strava.com/settings/api and change your “Authorization Callback Domain” to developers.strava.com. Please note, we only support Swagger 2.0. There is a known issue where you can only select one scope at a time. For more information, please check the section “client code” at https://developers.strava.com/docs.",
    "version" : "3.0.0"
  },
  "servers" : [ {
    "url" : "https://www.strava.com/api/v3"
  } ],
  "security" : [ {
    "strava_oauth" : [ "public" ]
  } ],
  "paths" : {
...
    "/clubs/{id}" : {
      "get" : {
        "tags" : [ "Clubs" ],
        "summary" : "Get Club",
        "description" : "Returns a given club using its identifier.",
        "operationId" : "getClubById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the club.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The detailed representation of a club.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DetailedClub"
                },
                "example" : {
                  "id" : 1,
                  "resource_state" : 3,
                  "name" : "Team Strava Cycling",
                  "profile_medium" : "https://dgalywyr863hv.cloudfront.net/pictures/clubs/1/1582/4/medium.jpg",
                  "profile" : "https://dgalywyr863hv.cloudfront.net/pictures/clubs/1/1582/4/large.jpg",
                  "cover_photo" : "https://dgalywyr863hv.cloudfront.net/pictures/clubs/1/4328276/1/large.jpg",
                  "cover_photo_small" : "https://dgalywyr863hv.cloudfront.net/pictures/clubs/1/4328276/1/small.jpg",
                  "sport_type" : "cycling",
                  "activity_types" : [ "Ride", "VirtualRide", "EBikeRide", "Velomobile", "Handcycle" ],
                  "city" : "San Francisco",
                  "state" : "California",
                  "country" : "United States",
                  "private" : true,
                  "member_count" : 116,
                  "featured" : false,
                  "verified" : false,
                  "url" : "team-strava-bike",
                  "membership" : "member",
                  "admin" : false,
                  "owner" : false,
                  "description" : "Private club for Cyclists who work at Strava.",
                  "club_type" : "company",
                  "post_count" : 29,
                  "owner_id" : 759,
                  "following_count" : 107
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/clubs/{id}/members" : {
      "get" : {
        "tags" : [ "Clubs" ],
        "summary" : "List Club Members",
        "description" : "Returns a list of the athletes who are members of a given club.",
        "operationId" : "getClubMembersById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the club.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number. Defaults to 1.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "per_page",
          "in" : "query",
          "description" : "Number of items per page. Defaults to 30.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 30
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A list of summary athlete representations.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SummaryAthlete"
                  },
                  "x-content-type" : "application/json"
                },
                "example" : [ {
                  "resource_state" : 2,
                  "firstname" : "Peter",
                  "lastname" : "S.",
                  "membership" : "member",
                  "admin" : false,
                  "owner" : false
                } ]
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/clubs/{id}/admins" : {
      "get" : {
        "tags" : [ "Clubs" ],
        "summary" : "List Club Administrators",
        "description" : "Returns a list of the administrators of a given club.",
        "operationId" : "getClubAdminsById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the club.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number. Defaults to 1.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "per_page",
          "in" : "query",
          "description" : "Number of items per page. Defaults to 30.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 30
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A list of summary athlete representations.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SummaryAthlete"
                  },
                  "x-content-type" : "application/json"
                },
                "example" : [ {
                  "resource_state" : 2,
                  "firstname" : "Peter",
                  "lastname" : "S."
                } ]
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/clubs/{id}/activities" : {
      "get" : {
        "tags" : [ "Clubs" ],
        "summary" : "List Club Activities",
        "description" : "Retrieve recent activities from members of a specific club. The authenticated athlete must belong to the requested club in order to hit this endpoint. Pagination is supported. Athlete profile visibility is respected for all activities.",
        "operationId" : "getClubActivitiesById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The identifier of the club.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number. Defaults to 1.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "per_page",
          "in" : "query",
          "description" : "Number of items per page. Defaults to 30.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 30
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A list of activities.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SummaryActivity"
                  },
                  "x-content-type" : "application/json"
                },
                "example" : [ {
                  "resource_state" : 2,
                  "athlete" : {
                    "resource_state" : 2,
                    "firstname" : "Peter",
                    "lastname" : "S."
                  },
                  "name" : "World Championship",
                  "distance" : 2641.7,
                  "moving_time" : 577,
                  "elapsed_time" : 635,
                  "total_elevation_gain" : 8.8,
                  "type" : "Ride",
                  "workout_type" : null
                } ]
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
    "/athlete/clubs" : {
      "get" : {
        "tags" : [ "Clubs" ],
        "summary" : "List Athlete Clubs",
        "description" : "Returns a list of the clubs whose membership includes the authenticated athlete.",
        "operationId" : "getLoggedInAthleteClubs",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "Page number. Defaults to 1.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "per_page",
          "in" : "query",
          "description" : "Number of items per page. Defaults to 30.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 30
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A list of summary club representations.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SummaryClub"
                  },
                  "x-content-type" : "application/json"
                },
                "example" : [ {
                  "id" : 231407,
                  "resource_state" : 2,
                  "name" : "The Strava Club",
                  "profile_medium" : "https://dgalywyr863hv.cloudfront.net/pictures/clubs/231407/5319085/1/medium.jpg",
                  "profile" : "https://dgalywyr863hv.cloudfront.net/pictures/clubs/231407/5319085/1/large.jpg",
                  "cover_photo" : "https://dgalywyr863hv.cloudfront.net/pictures/clubs/231407/5098428/4/large.jpg",
                  "cover_photo_small" : "https://dgalywyr863hv.cloudfront.net/pictures/clubs/231407/5098428/4/small.jpg",
                  "sport_type" : "other",
                  "city" : "San Francisco",
                  "state" : "California",
                  "country" : "United States",
                  "private" : false,
                  "member_count" : 93151,
                  "featured" : false,
                  "verified" : true,
                  "url" : "strava"
                } ]
              }
            }
          },
          "default" : {
            "description" : "Unexpected error.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fault"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Fault" : {
        "type" : "object",
        "properties" : {
          "errors" : {
            "type" : "array",
            "description" : "The set of specific errors associated with this fault, if any.",
            "items" : {
              "$ref" : "#/components/schemas/Error"
            }
          },
          "message" : {
            "type" : "string",
            "description" : "The message of the fault."
          }
        },
        "description" : "Encapsulates the errors that may be returned from the API."
      },
...
      "SummaryAthlete" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/MetaAthlete"
        }, {
          "type" : "object",
          "properties" : {
            "resource_state" : {
              "type" : "integer",
              "description" : "Resource state, indicates level of detail. Possible values: 1 -> \"meta\", 2 -> \"summary\", 3 -> \"detail\""
            },
            "firstname" : {
              "type" : "string",
              "description" : "The athlete's first name."
            },
            "lastname" : {
              "type" : "string",
              "description" : "The athlete's last name."
            },
            "profile_medium" : {
              "type" : "string",
              "description" : "URL to a 62x62 pixel profile picture."
            },
            "profile" : {
              "type" : "string",
              "description" : "URL to a 124x124 pixel profile picture."
            },
            "city" : {
              "type" : "string",
              "description" : "The athlete's city."
            },
            "state" : {
              "type" : "string",
              "description" : "The athlete's state or geographical region."
            },
            "country" : {
              "type" : "string",
              "description" : "The athlete's country."
            },
            "sex" : {
              "type" : "string",
              "description" : "The athlete's sex.",
              "enum" : [ "M", "F" ]
            },
            "premium" : {
              "type" : "boolean",
              "description" : "Deprecated.  Use summit field instead. Whether the athlete has any Summit subscription."
            },
            "summit" : {
              "type" : "boolean",
              "description" : "Whether the athlete has any Summit subscription."
            },
            "created_at" : {
              "type" : "string",
              "description" : "The time at which the athlete was created.",
              "format" : "date-time"
            },
            "updated_at" : {
              "type" : "string",
              "description" : "The time at which the athlete was last updated.",
              "format" : "date-time"
            }
          }
        } ]
      },
      "MetaAthlete" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "The unique identifier of the athlete",
            "format" : "int64"
          }
        },
        "example" : {
          "id" : 1
        }
      },
...
      "SummaryActivity" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/MetaActivity"
        }, {
          "type" : "object",
          "properties" : {
            "external_id" : {
              "type" : "string",
              "description" : "The identifier provided at upload time"
            },
            "upload_id" : {
              "type" : "integer",
              "description" : "The identifier of the upload that resulted in this activity",
              "format" : "int64"
            },
            "athlete" : {
              "$ref" : "#/components/schemas/MetaAthlete"
            },
            "name" : {
              "type" : "string",
              "description" : "The name of the activity"
            },
            "distance" : {
              "type" : "number",
              "description" : "The activity's distance, in meters",
              "format" : "float"
            },
            "moving_time" : {
              "type" : "integer",
              "description" : "The activity's moving time, in seconds"
            },
            "elapsed_time" : {
              "type" : "integer",
              "description" : "The activity's elapsed time, in seconds"
            },
            "total_elevation_gain" : {
              "type" : "number",
              "description" : "The activity's total elevation gain.",
              "format" : "float"
            },
            "elev_high" : {
              "type" : "number",
              "description" : "The activity's highest elevation, in meters",
              "format" : "float"
            },
            "elev_low" : {
              "type" : "number",
              "description" : "The activity's lowest elevation, in meters",
              "format" : "float"
            },
            "type" : {
              "$ref" : "#/components/schemas/ActivityType"
            },
            "start_date" : {
              "type" : "string",
              "description" : "The time at which the activity was started.",
              "format" : "date-time"
            },
            "start_date_local" : {
              "type" : "string",
              "description" : "The time at which the activity was started in the local timezone.",
              "format" : "date-time"
            },
            "timezone" : {
              "type" : "string",
              "description" : "The timezone of the activity"
            },
            "start_latlng" : {
              "$ref" : "#/components/schemas/LatLng"
            },
            "end_latlng" : {
              "$ref" : "#/components/schemas/LatLng"
            },
            "achievement_count" : {
              "type" : "integer",
              "description" : "The number of achievements gained during this activity"
            },
            "kudos_count" : {
              "type" : "integer",
              "description" : "The number of kudos given for this activity"
            },
            "comment_count" : {
              "type" : "integer",
              "description" : "The number of comments for this activity"
            },
            "athlete_count" : {
              "minimum" : 1,
              "type" : "integer",
              "description" : "The number of athletes for taking part in a group activity"
            },
            "photo_count" : {
              "type" : "integer",
              "description" : "The number of Instagram photos for this activity"
            },
            "total_photo_count" : {
              "type" : "integer",
              "description" : "The number of Instagram and Strava photos for this activity"
            },
            "map" : {
              "$ref" : "#/components/schemas/PolylineMap"
            },
            "trainer" : {
              "type" : "boolean",
              "description" : "Whether this activity was recorded on a training machine"
            },
            "commute" : {
              "type" : "boolean",
              "description" : "Whether this activity is a commute"
            },
            "manual" : {
              "type" : "boolean",
              "description" : "Whether this activity was created manually"
            },
            "private" : {
              "type" : "boolean",
              "description" : "Whether this activity is private"
            },
            "flagged" : {
              "type" : "boolean",
              "description" : "Whether this activity is flagged"
            },
            "workout_type" : {
              "type" : "integer",
              "description" : "The activity's workout type"
            },
            "upload_id_str" : {
              "type" : "string",
              "description" : "The unique identifier of the upload in string format"
            },
            "average_speed" : {
              "type" : "number",
              "description" : "The activity's average speed, in meters per second",
              "format" : "float"
            },
            "max_speed" : {
              "type" : "number",
              "description" : "The activity's max speed, in meters per second",
              "format" : "float"
            },
            "has_kudoed" : {
              "type" : "boolean",
              "description" : "Whether the logged-in athlete has kudoed this activity"
            },
            "hide_from_home" : {
              "type" : "boolean",
              "description" : "Whether the activity is muted"
            },
            "gear_id" : {
              "type" : "string",
              "description" : "The id of the gear for the activity"
            },
            "kilojoules" : {
              "type" : "number",
              "description" : "The total work done in kilojoules during this activity. Rides only",
              "format" : "float"
            },
            "average_watts" : {
              "type" : "number",
              "description" : "Average power output in watts during this activity. Rides only",
              "format" : "float"
            },
            "device_watts" : {
              "type" : "boolean",
              "description" : "Whether the watts are from a power meter, false if estimated"
            },
            "max_watts" : {
              "type" : "integer",
              "description" : "Rides with power meter data only"
            },
            "weighted_average_watts" : {
              "type" : "integer",
              "description" : "Similar to Normalized Power. Rides with power meter data only"
            }
          }
        } ]
      },
...
      "DetailedClub" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/SummaryClub"
        }, {
          "type" : "object",
          "properties" : {
            "membership" : {
              "type" : "string",
              "description" : "The membership status of the logged-in athlete.",
              "enum" : [ "member", "pending" ]
            },
            "admin" : {
              "type" : "boolean",
              "description" : "Whether the currently logged-in athlete is an administrator of this club."
            },
            "owner" : {
              "type" : "boolean",
              "description" : "Whether the currently logged-in athlete is the owner of this club."
            },
            "following_count" : {
              "type" : "integer",
              "description" : "The number of athletes in the club that the logged-in athlete follows."
            }
          }
        } ]
      },
...
    },
...
    "securitySchemes" : {
      "strava_oauth" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://www.strava.com/api/v3/oauth/authorize",
            "tokenUrl" : "https://www.strava.com/api/v3/oauth/token",
            "scopes" : {
              "read" : "Read public segments, public routes, public profile data, public posts, public events, club feeds, and leaderboards",
              "read_all" : "Read private routes, private segments, and private events for the user",
              "profile:read_all" : "Read all profile information even if the user has set their profile visibility to Followers or Only You",
              "profile:write" : "Update the user's weight and Functional Threshold Power (FTP), and access to star or unstar segments on their behalf",
              "activity:read" : "Read the user's activity data for activities that are visible to Everyone and Followers, excluding privacy zone data",
              "activity:read_all" : "The same access as activity:read, plus privacy zone data and access to read the user's activities with visibility set to Only You",
              "activity:write" : "Access to create manual activities and uploads, and access to edit any activities that are visible to the app, based on activity read access level"
            }
          }
        }
      }
    }
  },
...
}